Rename app key to user key to be consistent with the 3scale API
What we were calling 'app_key' is a 'user_key' in the 3scale API. It's
confusing because there is a different parameter for some of the calls in the
API called 'app_key'.

The problem could have been solved renaming the keys of the hashes that are
sent to the methods of the 3scale_client gem. However, I thought that using both
names through the codebase would be confusing.