authorizer: when limits are 0, return 'limits_exceeded' as the deny reason
We were returning 'disabled' when the usage limits for a metric were set to 0.
That was a requirement that we no longer have, so now we just return the
default error code returned by 3scale: 'limits_exceeded'.
In the future, we might return a more detailed reason (maybe including current
and max usage), so upper layers using this program, will be able to know
whether the limits are set to 0.