gemspec: update 3scale_client to v2.8.1
This new version of 3scale_client improves the parsing performance of the
response of authorize calls. That is relevant for the flusher as it performs
lots of authorize calls.
I have tried in local with docker containers (no latencies contacting 3scale)
to flush the reports generated for 100k combinations of {service, app, metric}
and I observed a 20% reduction in the time needed.