exe/xcflushd: do not specify default value for the 3scale port
The default value stopped applying when we started reading the options from
ENV. The problem is that gli uses the default value if it is not in the params,
but we only want to use ENV if the param is not present. When gli overwrites
it, we cannot know if it was specified by the user of filled by gli.