runner: shut down when receiving unhandled signals
This was already the case, but now we add a specific error log message.

There is currently no way in Ruby to set up sigprocmask, so while we
could go over the list of signals and ignore all of them, it is better
to cause termination on unknown signals than to ignore them at our own
peril. This is ok since the set of usual well known signals is already
being handled.

Note that not having sigprocmask available in Ruby might well be because
the VM is already providing guarantees when executing Ruby code.
Unfortunately this is all but well documented.