flusher, priority_auth_renewer: use a FixedThreadPool instead of a ThreadPoolExecutor
We are using concurrent-ruby's ThreadPoolExecutor without specifying a max size for the queue, and that type of pool only spawns a new thread when the queue is full.
That means that in practice, we specify a min:max number of threads, but only min are spawned. In the future, we might want to evaluate specifying a max size for the task queue, but for now, a FixedThreadPool is good enough.