priority_auth_renewer: wait less when publishing
We were waiting too much time between publishing attempts. The downside of this is that the thread pool quickly fills with tasks that are just waiting. With the new waiting times, the flusher processes around 5x the messages that it managed to process with the old configuration in my specific setup.
We might change this again in the future as we discover more use cases.