Dockerfile.verify: use root for accessing the volume
Sharing the project directory with a volume inside the container is
problematic when you want to use an unprivileged user. Apparently
Mac users get a fixed uid/gid combination, and Linux users get the
invoking user's ids. That is a can of worms that requires several
hacks and checks to work correctly. The current solution worked
because it happened to be a coincidence in the uids for single user
machines (ie. uid 1001), but it breaks as soon as you have another
uid invoke it.

TL;DR: just use root inside the container to avoid EPERM.